Introduce TST_LIST variable to select a test to execute
It might be desirable, due to restrictions in the testing environment, to
execute tests individually. This patch introduces the TST_LIST variable,
which can be set with the name of the test to execute. If the variable is
set, expect_pass and expect_fail automatically skip the tests when the
first argument of those functions does not match the value of TST_LIST.

TST_LIST can be also used in conjunction with the UML kernel. It is
sufficient to add it to the kernel command line.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>